requirements.txt
google-genai==1.12.1
python-dotenv==1.1.0
docopt==0.6.2
